+++ stable/index.js
@@ -13,7 +13,7 @@ function pluginYummyAnime() {
 
     window.LampaYani = window.LampaYani || {};
     window.LampaYani.Config = window.LampaYaniConfig = {
-        version: '0.44.6',
+        version: '0.44.7',
         apiBase: 'https://api.yani.tv',
         statusUrl: 'https://yummyanime.github.io/yummy-lampa-plugin/status/status.json',
         applicationHeader: defaultApplicationToken, // Backward-compatible default public token.
@@ -8077,6 +8077,8 @@ function pluginYummyAnime() {
 
     function create(object, deps) {
         var t = deps.t, locale = deps.locale, toCard = deps.toCard;
+        var playbackSourceId = deps.playbackSourceId || function () { return ''; };
+        var isPlaybackSourceEnabled = deps.isPlaybackSourceEnabled || function () { return true; };
         var scroll = new Lampa.Scroll({mask: true, over: true, step: 250});
         scroll.minus();
         var html = $('<div class="yani-schedule"></div>');
@@ -8147,6 +8149,8 @@ function pluginYummyAnime() {
                 var data = window.LampaYaniUiUtils && LampaYaniUiUtils.videoData
                     ? LampaYaniUiUtils.videoData(video)
                     : video && video.data || {};
+                var sourceId = playbackSourceId(Object.assign({}, video || {}, data || {}));
+                if (!isPlaybackSourceEnabled(sourceId)) return;
                 rememberTranslationName(bucket, data.dubbing || data.translation || data.voice || video.dub_title || video.dubbing);
             });
             return {
@@ -15012,6 +15016,8 @@ function pluginYummyAnime() {
             t: t,
             locale: locale,
             toCard: toCard,
+            playbackSourceId: playbackSourceId,
+            isPlaybackSourceEnabled: isPlaybackSourceEnabled,
             openYummyDetail: openYummyDetail,
             goBack: goBack
         });
@@ -16788,26 +16794,6 @@ function pluginYummyAnime() {
                 onChange: openSettingsLogin
             });
         }
-
-        Lampa.SettingsApi.addParam({
-            component: 'yani',
-            param: {name: 'yani_api_settings_title', type: 'title'},
-            field: {name: t('api_settings')}
-        });
-
-        Lampa.SettingsApi.addParam({
-            component: 'yani',
-            param: {name: 'yani_api_check', type: 'button'},
-            field: {name: t('api_check_name'), description: t('api_check_description')},
-            onChange: function () {
-                LampaYaniApi.health().then(function () {
-                    Lampa.Noty.show(t('api_ok'));
-                }).catch(function (error) {
-                    console.error('[YummyAnime]', error);
-                    Lampa.Noty.show(t('api_error'));
-                });
-            }
-        });
 
         Lampa.SettingsApi.addParam({
             component: 'yani',